requirements.txt
streamlit
streamlit-drawable-canvas
langchain
langchain-google-genai
python-dotenv
pillow
